[Refactor/#39] 피드백 모두 반영
