+++ src/view/main/EditTodo/editTodoCard.css
@@ -1,0 +1,44 @@
+.edit-wrap {
+  min-width: 200px;
+  max-width: 100%;
+  display: flex;
+  flex-direction: column;
+
+  gap: 12px;
+}
+
+.edit-title {
+  margin-top: 10px;
+  padding: 6px;
+
+  font-size: 1rem;
+}
+
+.edit-date {
+  padding: 6px;
+
+  font-size: 1rem;
+}
+
+.edit-wrap > textarea {
+  padding: 6px;
+
+  font-size: 1rem;
+}
+
+.edit-btn {
+  padding: 0px 26px;
+
+  border: 0.1rem solid rgba(99, 99, 99, 0.3);
+  border-radius: 10px;
+
+  font-size: 1rem;
+  background-color: #fff;
+
+  cursor: cell;
+}
+
+.edit-btn:hover {
+  background-color: rgba(99, 99, 99, 0.3);
+  border: 1px solid #fff;
+}
