--- src/view/main/todoCardList/todoCardList.css
@@ -1,13 +1,0 @@
-@charset "utf-8";
-
-.todo-card-wrap {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-  grid-auto-rows: auto;
-
-  gap: 40px;
-}
-
-.todo-card {
-  grid-row: span min-content;
-}
