+++ src/view/main/todolist/todolist.css
@@ -3,5 +3,112 @@
 main {
   display: flex;
   flex-direction: column;
+  gap: 30px;
+}
+.todo-card-wrap {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+  grid-auto-rows: auto;
+
   gap: 40px;
+}
+
+.todo-card {
+  grid-row: span min-content;
+}
+
+.todo-card h4 {
+  font-size: 1.2rem;
+  margin: 1rem 0 1rem 0;
+}
+.work-nd-done {
+  padding: 20px;
+
+  border-radius: 20px;
+  box-shadow: 0px 2px 5px 1px rgba(0, 0, 0, 0.2);
+}
+.view-comment {
+  word-break: break-all;
+  font-size: 1rem;
+}
+
+.view-date {
+  width: 100%;
+
+  margin-top: 0;
+  margin-bottom: 1rem;
+
+  font-size: 1rem;
+  text-align: end;
+}
+
+/* 삭제, 완료 / 취소, 저장버튼 감싸는 틀 */
+.btn-delete-nd-done,
+.btn-cancel-nd-save {
+  width: 100%;
+
+  display: flex;
+  justify-content: space-between;
+}
+
+.delete-btn {
+  background-color: #feb198;
+  padding: 8px 10px;
+
+  border: none;
+  border-radius: 10px;
+
+  font-size: 1rem;
+  cursor: cell;
+}
+
+.delete-btn:hover,
+.delete-btn:focus,
+.cancel-btn:hover {
+  background-color: rgba(254, 177, 152, 0.6);
+}
+
+.cancel-btn {
+  background-color: #feb198;
+  padding: 8px 26px;
+
+  border: none;
+  border-radius: 10px;
+
+  font-size: 1rem;
+  cursor: cell;
+}
+
+/* 완료버튼과 저장버튼 */
+.done-btn {
+  padding: 8px 26px;
+
+  border: none;
+  border-radius: 10px;
+
+  font-size: 1rem;
+  background-color: #333;
+  color: white;
+
+  cursor: cell;
+}
+
+.done-btn:hover,
+.done-btn:focus,
+.save-btn:hover {
+  background-color: rgba(128, 128, 128, 0.5);
+  color: #111;
+}
+
+.save-btn {
+  padding: 8px 26px;
+
+  border: none;
+  border-radius: 10px;
+
+  font-size: 1rem;
+  background-color: #333;
+  color: white;
+
+  cursor: cell;
 }
